% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{bind_extracted_data}
\alias{bind_extracted_data}
\title{Bind dataframes with same names from a list that contains lists of dataframes}
\usage{
bind_extracted_data(result_list)
}
\description{
Bind dataframes with same names from a list that contains lists of dataframes
}
