% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{check_extracted_data_fields}
\alias{check_extracted_data_fields}
\title{Check if extraction of data captured all available fields}
\usage{
check_extracted_data_fields(data_entry, extracted_data, warnings)
}
\description{
Check if extraction of data captured all available fields
}
