% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{concat}
\alias{concat}
\alias{or}
\alias{and}
\alias{includes}
\title{Concatenate dimensions search filters}
\usage{
concat(x, sep1 = c(" and ", " or "), sep2 = c(" and ", " or "))

or(x)

and(x)

includes(x, sep1 = c(" and ", " or "), sep2 = c(" and ", " or "))
}
\description{
Concatenate dimensions search filters
}
\section{Functions}{
\itemize{
\item \code{or}: Creates filter with ' or '.

\item \code{and}: Creates filter with ' and '.

\item \code{includes}: Creates filter with 'in [...]'.
}}

