% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_helper_functions.R
\name{create_graph_from_bipartite_data}
\alias{create_graph_from_bipartite_data}
\title{Create one-mode igraph object from bipartite network data.}
\usage{
create_graph_from_bipartite_data(
  edge_list,
  node_attributes = NULL,
  node_id = "researcher_id",
  projection_to_extract = 2
)
}
\description{
Create one-mode igraph object from bipartite network data.
}
