% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_author_field_of_research.R
\name{create_legend_color_df}
\alias{create_legend_color_df}
\title{Create legend_color_df to use with `legend()`.}
\usage{
create_legend_color_df(authors_research_colors, fields_to_color_count = 9)
}
\description{
Create legend_color_df to use with `legend()`.
}
