% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fields.R
\name{determine_field_type}
\alias{determine_field_type}
\title{Cycles through all results and determines the type/class of for a specified data field.}
\usage{
determine_field_type(raw_elements, field_name)
}
\description{
Cycles through all results and determines the type/class of for a specified data field.
}
