% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_researchers.R
\name{dimensions_id_finder}
\alias{dimensions_id_finder}
\title{Web Interface for finding researchers' dimensions ID}
\usage{
dimensions_id_finder()
}
\description{
Web Interface for finding researchers' dimensions ID
}
