% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions_auth.R
\name{dimensions_login}
\alias{dimensions_login}
\title{Login to the Dimensions Analytics API}
\usage{
dimensions_login(credentials = NULL)
}
\arguments{
\item{credentials}{(list) List of user credentials, including username
(email) and password. To avoid providing a manual list each time, you may
instead add your credentials to your .Renviron file. These should be named
\code{dimensions_username} and \code{dimensions_password}. \code{dimensions_login()} will
then use these credentials automatically to login.}
}
\description{
Login to the Dimensions Analytics API
}
\examples{
# Login with a list of credentials
dimensions_login(credentials = list(
    "username" = "your_username",
    "password" = "your_password"))

Alternatively with credentials added to .Renviron file:
dimensions_login()
}
