% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions_auth.R
\name{dimensions_logout}
\alias{dimensions_logout}
\title{Logout of the Dimensions Analytics API}
\usage{
dimensions_logout()
}
\description{
Logout of the Dimensions Analytics API
}
\examples{
# Logout and destroy existing login token
dimensions_logout()
}
