% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions_query.R
\name{dimensions_query}
\alias{dimensions_query}
\title{Make a basic query request to the Dimensions Analytics API}
\usage{
dimensions_query(query = NULL, format = "list")
}
\arguments{
\item{query}{(character) String containing a raw Dimensions Search Language
query.}

\item{format}{(character) The format in which data should be returned. One of
"list" or "json"}
}
\description{
Make a basic query request to the Dimensions Analytics API
}
\examples{
# The most basic query: search and return all publications
dimensions_query(query = "search publications return publications")

# Return results in JSON format
dimensions_query(query = "search publications return publications",
format = "json")

# Dimensions automatically limits results to 20 records. Add a "limit" clause
to increase the number of returned results
dimensions_query(query = "search publications return publications limit 100")

# Search for a specific DOI. Note that quotation marks surrounding DOIs must
be escaped by placing a backwards slash (\) in front of the quotation marks.
You may find the `paste` and `paste0` functions helpful to build longer query
strings.
my_query <- paste0("search publications",
                   "where doi = \"10.3389/frma.2018.00023\"",
                   "return publications")
dimensions_query(query = my_query)



A full overview of query syntax can be found in the [Dimensions Search
Language](https://docs.dimensions.ai/dsl/) documentation.
}
