% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions_raw.R
\name{dimensions_raw}
\alias{dimensions_raw}
\title{Make a raw query request to the Dimensions Analytics API}
\usage{
dimensions_raw(query = NULL, format = "list")
}
\arguments{
\item{query}{(character) String containing a raw Dimensions Search Language
query.}

\item{format}{(character) The format in which data should be returned. One of
"list" or "json"}
}
\description{
Make a raw query request to the Dimensions Analytics API
}
\examples{
# The most basic query: search and return all publications
dimensions_raw(query = "search publications return publications")

# Dimensions automatically limits results to 20 records. Add a "limit" clause
to increase the number of returned results
dimensions_raw(query = "search publications return publications limit 100")

# Search for a specific DOI. Note that quotation marks surrounding DOIs must
be escaped by placing a backwards slash (\) in front of the quotation marks.
You may find the `paste0` function helpful to build longer query strings.
dimensions_raw(query = paste0("search publications",
                              "where doi = \"10.3389/frma.2018.00023\"",
                              "return publications")

A full overview of query syntax can be found in the [Dimensions Search
Language](https://docs.dimensions.ai/dsl/) documentation.
}
