% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions_request.R
\name{dimensions_request}
\alias{dimensions_request}
\title{Send DSL query to Dimensions API and collect results as list based on json structure}
\usage{
dimensions_request(
  dsl_query = "describe",
  session_token = getOption("dimensions_token"),
  fail_on_error = TRUE
)
}
\description{
Send DSL query to Dimensions API and collect results as list based on json structure
}
