% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{dimensions_search}
\alias{dimensions_search}
\title{Search dimensions API and compose paginated query text}
\usage{
dimensions_search(
  search_source = c("publications", "grants", "patents", "clinical_trials",
    "policy_documents", "researchers", "organizations", "datasets"),
  search_term = NULL,
  search_filter = NULL,
  return = search_source,
  return_fields = "all",
  sort_by = NULL,
  paginate = TRUE,
  page_length = 1000,
  print_query = FALSE,
  session_token = getOption("dimensions_token")
)
}
\description{
Search dimensions API and compose paginated query text
}
