% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{extract_data}
\alias{extract_data}
\title{Extract data from a single dimensions request result list into a dataframe}
\usage{
extract_data(data_entry, data_source, deprecated_fields = NULL)
}
\description{
Extract data from a single dimensions request result list into a dataframe
}
