% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_department.R
\name{extract_department_name}
\alias{extract_department_name}
\title{Extract the department affiliation from raw_affiliations}
\usage{
extract_department_name(raw_aff)
}
\arguments{
\item{raw_aff}{Vector of raw affiliation characters.}
}
\description{
Extract the department affiliation from raw_affiliations
}
