% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fields.R
\name{extract_field}
\alias{extract_field}
\title{Extracts one field from a list of `raw_elements`.}
\usage{
extract_field(raw_elements, field_name, atomic, has_names, id, null_lgl)
}
\description{
Extracts one field from a list of `raw_elements`.
}
