% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fields.R
\name{extract_fields}
\alias{extract_fields}
\title{Extracts all data fields from a `raw_elements` list.}
\usage{
extract_fields(raw_elements)
}
\arguments{
\item{raw_elements}{`List` of results, as returned by `dimensions_request()`}
}
\description{
`dimensions_request(..., auto.extract = FALSE)` returns a
nested list. `extract_fields()` can be used to turn that list
into a list of `data.frames` where each field is represented by
`one data.frame`.
}
