% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_bi_graph.R
\name{fast_bi_igraph}
\alias{fast_bi_igraph}
\title{Fastly create igraph object from bipartite data using data.table}
\usage{
fast_bi_igraph(data, edge_threshold = 1300, skip_make_unique = FALSE)
}
\arguments{
\item{data}{`data.frame` with from and to column}

\item{edge_threshold}{`numeric` defining minimum edges for an edge to be included}

\item{skip_make_unique}{`logical`, set `TRUE` to include duplicated rows in `data`.}
}
\description{
Fastly create igraph object from bipartite data using data.table
}
