% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_author_field_of_research.R
\name{get_author_field_of_research}
\alias{get_author_field_of_research}
\title{Extract most frequent field of research (FoR) an author published with.}
\usage{
get_author_field_of_research(category_for, authors)
}
\description{
Extract most frequent field of research (FoR) an author published with.
}
