% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_descr.R
\name{get_data_descriptions}
\alias{get_data_descriptions}
\title{Get descriptions of all dimensions.ai data sources}
\usage{
get_data_descriptions(..., tidy_output = TRUE)
}
\arguments{
\item{token}{Character containg a dimensions.ai API token, use get_token() to aquire token.}
}
\description{
Get descriptions of all dimensions.ai data sources
}
