% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_deprecated_fields.R
\name{get_deprecated_fields}
\alias{get_deprecated_fields}
\title{Get deprecated fields for dimensions.ai data sources}
\usage{
get_deprecated_fields(
  data_source = c("all", "publications", "researchers", "patents", "policy_documents",
    "clinical_trials", "grants")
)
}
\arguments{
\item{data_source}{Character specifying data source, defaults to all.}
}
\description{
Get deprecated fields for dimensions.ai data sources
}
