% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_descr.R
\name{get_source_descriptions}
\alias{get_source_descriptions}
\title{Get descriptions of all dimensions.ai data sources}
\usage{
get_source_descriptions(..., tidy_output = TRUE)
}
\arguments{
\item{token}{Character containing a dimensions.ai API token, use get_token() to acquire token.}
}
\description{
Get descriptions of all dimensions.ai data sources
}
