% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{get_token}
\alias{get_token}
\title{Get a dimensions.ai API token}
\usage{
get_token(login)
}
\arguments{
\item{login}{A list with named entries "username" and "password"}
}
\description{
Get a dimensions.ai API token
}
