% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_department.R
\name{group_departments}
\alias{group_departments}
\title{Map/ group departments to/by college or other grouping}
\usage{
group_departments(
  authors_df,
  department_name_list,
  department_var = "department"
)
}
\arguments{
\item{authors_df}{Author dataframe extracted from dimensioans.ai data.}

\item{department_name_list}{List of department names. This should NOT include phrases like "Deparment of...", just the names of the disciplines.}

\item{deparment_var}{Name of the variable that contains the department affiliation.}
}
\description{
Map/ group departments to/by college or other grouping
}
