% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_author_field_of_research.R
\name{mutate_authors_field_of_research_to_colors}
\alias{mutate_authors_field_of_research_to_colors}
\title{Mutate author_field_of_research to contain colors for biggest fields.}
\usage{
mutate_authors_field_of_research_to_colors(
  authors_field_of_research,
  fields_to_color_count = 9
)
}
\description{
Mutate author_field_of_research to contain colors for biggest fields.
}
