% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paginate_query_text.R
\name{paginate_query_text}
\alias{paginate_query_text}
\alias{split_query_text}
\alias{split_and_replace_query_text}
\title{Create a pagination character vector to add to a query with more than 1000 results}
\usage{
paginate_query_text(
  query_text,
  page_length = 1000,
  session_token = getOption("dimensions_token")
)

split_query_text(query_texts, query_number, page_length = 100)

split_and_replace_query_text(query_texts, query_number, page_length = 100)
}
\description{
Create a pagination character vector to add to a query with more than 1000 results
}
\section{Functions}{
\itemize{
\item \code{split_query_text}: Split query text entry in a vector previously created with paginate_query_text into smaller pages. Useful, when any query text fails to complete and you want to rerun only those that failed.

\item \code{split_and_replace_query_text}: Like `split_query_text()` but returns full list of query text including unchanged ones. Useful, when any query text fails to complete and you want to rerun all queries.
}}

