% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_helper_functions.R
\name{plot_graph_with_legend_and_stats}
\alias{plot_graph_with_legend_and_stats}
\title{Plot graph with stress layout and optionally with legend and some statistics.}
\usage{
plot_graph_with_legend_and_stats(
  g,
  layout = NULL,
  stats = NULL,
  legend_color_df = NULL,
  main = NULL
)
}
\description{
Plot graph with stress layout and optionally with legend and some statistics.
}
