% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_to_db.R
\name{table_list_to_db}
\alias{table_list_to_db}
\title{Store list of dataframes to database}
\usage{
table_list_to_db(table_list, data_source, db_connection)
}
\description{
Store list of dataframes to database
}
